+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/combat/fletching/entity/projectile/TorchArrow.java
@@ -6,11 +6,13 @@ import net.minecraft.core.Direction;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.Arrow;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.StandingAndWallBlockItem;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.EntityHitResult;
@@ -36,15 +38,11 @@ public class TorchArrow extends Arrow {
         }
         Direction direction = pResult.getDirection();
         BlockPos pos = pResult.getBlockPos().relative(direction);
-        BlockState hitState = this.level().getBlockState(pos);
-        BlockState stateToPlace = null;
-        if (direction != Direction.UP) {
-            BlockPlaceContext blockPlaceContext = new BlockPlaceContext(this.level(), null, InteractionHand.MAIN_HAND, getPickupItem(), pResult);
-            stateToPlace = Blocks.WALL_TORCH.getStateForPlacement(blockPlaceContext);
-        }
-        else if (hitState.canBeReplaced()) {
-            stateToPlace = Blocks.TORCH.defaultBlockState();
-        }
+        Player player = null;
+        if (this.getOwner() instanceof Player)
+            player = (Player) this.getOwner();
+        BlockPlaceContext blockPlaceContext = new BlockPlaceContext(this.level(), player, InteractionHand.MAIN_HAND, getPickupItem(), pResult);
+        BlockState stateToPlace = ((StandingAndWallBlockItem)Items.TORCH).getPlacementState(blockPlaceContext);
 
         if (stateToPlace != null) {
             this.level().setBlock(pos, stateToPlace, 3);